Cast nparams to int when checking if it's >= 0

+++ userspace/libscap/scap_savefile.c
@@ -2826,7 +2826,7 @@ int32_t scap_next_offline(scap_t *handle, OUT scap_evt **pevent, OUT uint16_t *p
 			uint16_t *lens = (uint16_t *)((char *)*pevent + sizeof(struct ppm_evt_hdr));
 			uint32_t nparams;
 			bool done = false;
-			for(nparams = g_event_info[(*pevent)->type].nparams; nparams >= 0; nparams--)
+			for(nparams = g_event_info[(*pevent)->type].nparams; (int)nparams >= 0; nparams--)
 			{
 				char *valptr = (char *)lens + nparams * sizeof(uint16_t);
 				if(valptr > end)
